add auth

+++ src/main/java/org/elcer/accounts/app/SampleDataInitializer.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
+import java.math.BigDecimal;
 
 @Service
 public class SampleDataInitializer {
@@ -23,7 +24,7 @@ public class SampleDataInitializer {
         if (init) return;
         try {
             for (int i = 1; i < ACCOUNS_TO_CREATE; i++) {
-                Account account = new Account(RandomUtils.getGtZeroRandom());
+                Account account = new Account(BigDecimal.valueOf(RandomUtils.getGtZeroRandom()));
                 accountRepository.save(account);
             }
         } finally {
